+++ gherkin/javascript/src/index.ts
@@ -1,4 +1,6 @@
 import GherkinStreams from './stream/GherkinStreams'
+import generateMessages from './stream/generateMessages'
+import makeSourceEnvelope from './stream/makeSourceEnvelope'
 import IGherkinOptions from './IGherkinOptions'
 import Query from './Query'
 import Dialect from './Dialect'
@@ -7,12 +9,13 @@ import AstBuilder from './AstBuilder'
 import TokenScanner from './TokenScanner'
 import compile from './pickles/compile'
 import DIALECTS from './gherkin-languages.json'
-import { parse, parseAndCompile } from '../test/utils/parse'
 
 const dialects = DIALECTS as Readonly<{ [key: string]: Dialect }>
 
 export {
   GherkinStreams,
+  generateMessages,
+  makeSourceEnvelope,
   IGherkinOptions,
   Query,
   dialects,
@@ -21,6 +24,4 @@ export {
   AstBuilder,
   TokenScanner,
   compile,
-  parse,
-  parseAndCompile,
 }