+++ react/javascript/test/search/TextSearchTest.ts
@@ -1,7 +1,8 @@
 import assert from 'assert'
 import TextSearch from '../../src/search/TextSearch'
-import { parse } from '@cucumber/gherkin'
 import { pretty } from '@cucumber/gherkin-utils'
+import { IdGenerator, messages } from '@cucumber/messages'
+import { AstBuilder, Parser } from '@cucumber/gherkin'
 
 describe('TextSearch', () => {
   let search: TextSearch
@@ -135,3 +136,11 @@ describe('TextSearch', () => {
     })
   })
 })
+
+function parse(source: string): messages.IGherkinDocument {
+  const newId = IdGenerator.uuid()
+  const parser = new Parser(new AstBuilder(newId))
+  const gherkinDocument = parser.parse(source)
+  gherkinDocument.uri = ''
+  return gherkinDocument
+}